CR-935 remove duplicate firestore classes in favour of the cucumber-common wrapper
