+++ src/test/java/uk/gov/ons/ctp/integration/contcencucumber/main/SpringIntegrationTest.java
@@ -8,16 +8,16 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.web.client.RestTemplate;
-import uk.gov.ons.ctp.integration.contcencucumber.cloud.CloudDataStore;
-import uk.gov.ons.ctp.integration.contcencucumber.main.repository.CaseDataRepository;
+import uk.gov.ons.ctp.common.cloud.FirestoreDataStore;
+import uk.gov.ons.ctp.common.cloud.TestCloudDataStore;
 import uk.gov.ons.ctp.integration.contcencucumber.main.service.ProductService;
 
 @ContextConfiguration(
     classes = {
       SpringIntegrationTest.class,
       ProductService.class,
-      CaseDataRepository.class,
-      CloudDataStore.class
+      TestCloudDataStore.class,
+      FirestoreDataStore.class
     },
     loader = SpringBootContextLoader.class,
     initializers = ConfigFileApplicationContextInitializer.class)
@@ -26,8 +26,7 @@ import uk.gov.ons.ctp.integration.contcencucumber.main.service.ProductService;
 @Import({
   uk.gov.ons.ctp.integration.contcencucumber.main.service.impl.ProductServiceImpl.class,
   uk.gov.ons.ctp.integration.common.product.ProductReference.class,
-  uk.gov.ons.ctp.integration.contcencucumber.main.repository.impl.CaseDataRepositoryImpl.class,
-  uk.gov.ons.ctp.integration.contcencucumber.cloud.FirestoreDataStore.class
+  uk.gov.ons.ctp.integration.contcencucumber.main.repository.CaseDataRepository.class
 })
 public class SpringIntegrationTest {
