+++ src/test/java/uk/gov/ons/ctp/integration/contcencucumber/main/repository/CaseDataRepository.java
@@ -1,28 +1,41 @@
 package uk.gov.ons.ctp.integration.contcencucumber.main.repository;
 
 import java.util.List;
+import javax.annotation.PostConstruct;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import uk.gov.ons.ctp.common.cloud.TestCloudDataStore;
 import uk.gov.ons.ctp.common.domain.UniquePropertyReferenceNumber;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.integration.contcencucumber.cloud.CachedCase;
 
-/** Repository for Case Data */
-public interface CaseDataRepository {
+@Service
+public class CaseDataRepository {
+  private static final String[] SEARCH_BY_UPRN_PATH = new String[] {"uprn"};
 
-  /**
-   * Get all Cached cases for an address by Unique Property Reference Number.
-   *
-   * @param uprn UPRN of the case to read
-   * @return list of cached cases found that match the given UPRN
-   * @throws CTPException on error
-   */
-  List<CachedCase> readCachedCasesByUprn(final UniquePropertyReferenceNumber uprn)
-      throws CTPException;
+  @Value("${google-cloud-project}")
+  private String gcpProject;
 
-  /**
-   * Delete a case from Firestore
-   *
-   * @param key of case to delete
-   * @throws CTPException
-   */
-  void deleteCachedCase(String key) throws CTPException;
+  @Value("${cloud-storage.case-schema-name}")
+  private String caseSchemaName;
+
+  private String caseSchema;
+
+  @Autowired private TestCloudDataStore cloudDataStore;
+
+  @PostConstruct
+  public void init() {
+    caseSchema = gcpProject + "-" + caseSchemaName.toLowerCase();
+  }
+
+  public List<CachedCase> readCachedCasesByUprn(UniquePropertyReferenceNumber uprn)
+      throws CTPException {
+    String key = String.valueOf(uprn.getValue());
+    return cloudDataStore.search(CachedCase.class, caseSchema, SEARCH_BY_UPRN_PATH, key);
+  }
+
+  public void deleteCachedCase(String key) throws CTPException {
+    cloudDataStore.deleteObject(caseSchema, key);
+  }
 }
